move cartSchema and goodsSchema to model
