+++ router/route/cart.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 
-var cartSchema = require('./module/cartSchema.js');
+var cartSchema = require('./model/cartSchema.js');
 
 router.get('/', function (req, res, next) {
   cartSchema.find(function (err, cartGoods) {
