+++ router/route/shop.js
@@ -1,7 +1,7 @@
 var express = require('express');
 var router = express.Router();
 var mongoose = require('mongoose');
-var connect = require('./module/goodsSchema.js');
+var connect = require('./model/goodsSchema.js');
 
 router.get('/', function (req, res, next) {
   connect.find(function (err, goods) {
