+++ router/route/shopManagement.js
@@ -2,7 +2,7 @@ var express = require('express');
 var router = express.Router();
 
 var mongoose = require('mongoose');
-var connect = require('./module/goodsSchema.js');
+var connect = require('./model/goodsSchema.js');
 
 router.get('/', function (req, res, next) {
